# openapi.yaml
openapi: 3.1.0
info:
  title: Silicon Briefcase API
  version: 1.1.0
  description: Organization-scoped file storage, permissions, previews, search, versioning, recovery, and OBO application
    access.
servers:
- url: https://backend.briefcase.teamofsilicons.com/api/v1
  description: The API is served by the backend host. Permanent entry URLs are minted on the application host instead,
    always organization-scoped as https://briefcase.teamofsilicons.com/org/{org_id}/{path}.
security:
- bearerAuth: []
- bearerAuth: []
  testingEnvironmentKey: []
...
paths:
...
  /entries/{entry_id}/permissions:
    parameters:
    - $ref: '#/components/parameters/OrgId'
    - $ref: '#/components/parameters/EntryId'
    get:
      tags:
      - Permissions
      operationId: listPermissions
      responses:
        '200':
          description: Explicit permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/PermissionGrant'
        default:
          $ref: '#/components/responses/Error'
      x-operation-version: 1.0.0
    post:
      tags:
      - Permissions
      operationId: grantPermission
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermissionGrantCreate'
      responses:
        '201':
          description: Permission granted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermissionGrant'
        default:
          $ref: '#/components/responses/Error'
      x-operation-version: 1.0.0
  /entries/{entry_id}/permissions/{grant_id}:
    delete:
      tags:
      - Permissions
      operationId: revokePermission
      parameters:
      - $ref: '#/components/parameters/OrgId'
      - $ref: '#/components/parameters/EntryId'
      - name: grant_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Grant revoked
        default:
          $ref: '#/components/responses/Error'
      x-operation-version: 1.0.0
...
  /obo/invitations:
    post:
      operationId: inviteOnBehalfOfMember
      summary: Critical IAM-approved operation inside the calling app subtree
      tags:
      - Permissions
      parameters:
      - $ref: '#/components/parameters/OboOrgId'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invitation'
        default:
          $ref: '#/components/responses/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DelegatedInvite'
      security:
      - oboAccess: []
        appId: []
      - oboAccess: []
        appId: []
        testingEnvironmentKey: []
      x-iam-endpoint-id: briefcase.invitations.create
      x-iam-critical: true
      description: IAM binds POST, this exact versioned path, the exact body SHA-256, subject, organization, and
        calling application. Require critical user approval in the IAM endpoint catalog.
      x-operation-version: 1.0.0
  /obo/link-access:
    post:
      operationId: setLinkAccessOnBehalfOfMember
      summary: Critical IAM-approved operation inside the calling app subtree
      tags:
      - Permissions
      parameters:
      - $ref: '#/components/parameters/OboOrgId'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkAccess'
        default:
          $ref: '#/components/responses/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DelegatedLinkAccess'
      security:
      - oboAccess: []
        appId: []
      - oboAccess: []
        appId: []
        testingEnvironmentKey: []
      x-iam-endpoint-id: briefcase.link_access.update
      x-iam-critical: true
      description: IAM binds POST, this exact versioned path, the exact body SHA-256, subject, organization, and
        calling application. Require critical user approval in the IAM endpoint catalog.
      x-operation-version: 1.1.0
components:
  securitySchemes:
    uploadCapability:
      type: apiKey
      in: header
      name: X-Briefcase-Upload-Capability
      description: Narrow private-byte capability issued by an authorized upload reservation
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: IAM access token
    oboAccess:
      type: apiKey
      in: header
      name: X-IAM-OBO-Access-Proof
      description: Short-lived IAM OBO Access proof
    appId:
      type: apiKey
      in: header
      name: X-App-ID
      description: Originating application identifier required together with an OBO Access proof
    testingEnvironmentKey:
      type: apiKey
      in: header
      name: X-Briefcase-App-Secret
      description: 'IAM test Application secret: ask_ followed by 43 URL-safe characters. Selects an already provisioned
        and registered Briefcase testing environment.'
  parameters:
    OrgId:
      name: X-Org-ID
      in: header
      required: true
      schema:
        type: string
    OboOrgId:
      name: X-Org-ID
      in: header
      required: false
      schema:
        type: string
      description: Optional; IAM derives the organization and any supplied value must agree
...
    EntryId:
      name: entry_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    ActorRef:
      type: object
      required:
      - type
      - id
      properties:
        type:
          type: string
          enum:
          - carbon
          - silicon
          - application
        id:
          type: string
...
    PermissionGrantCreate:
      type: object
      required:
      - principal
      - access
      properties:
        principal:
          $ref: '#/components/schemas/ActorRef'
        access:
          type: array
          minItems: 1
          items:
            type: string
            enum:
            - read
            - write
            - update
        inherit:
          type: boolean
          default: true
    PermissionGrant:
      allOf:
      - $ref: '#/components/schemas/PermissionGrantCreate'
      - type: object
        required:
        - id
        - granted_by
        - created_at
        properties:
          id:
            type: string
            format: uuid
          granted_by:
            $ref: '#/components/schemas/ActorRef'
          created_at:
            type: string
            format: date-time
...
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
...
    InvitationRecipient:
      type: object
      additionalProperties: false
      required:
      - type
      - id
      properties:
        type:
          type: string
          enum:
          - carbon
          - silicon
          - email
          - tag
        id:
          type: string
          minLength: 1
          maxLength: 254
    InvitationCreate:
      type: object
      additionalProperties: false
      required:
      - principal
      properties:
        principal:
          $ref: '#/components/schemas/InvitationRecipient'
        access:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            enum:
            - read
            - write
            - update
          default:
          - read
        inherit:
          type: boolean
          default: true
    Invitation:
      type: object
      additionalProperties: false
      required:
      - id
      - principal
      - access
      - inherit
      properties:
        id:
          type: string
          format: uuid
        principal:
          $ref: '#/components/schemas/InvitationRecipient'
        access:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            enum:
            - read
            - write
            - update
          default:
          - read
        inherit:
          type: boolean
          default: true
...
    LinkAccess:
      type: object
      additionalProperties: false
      required:
      - enabled
      - effective
      - inherited_from
      - can_manage
      - url
      properties:
        url:
          type:
          - string
          - 'null'
          format: uri
          description: Shareable file or folder website URL when effective is true; otherwise null. Testing URLs require the recipient to select the same testing environment.
        enabled:
          type: boolean
        effective:
          type: boolean
        inherited_from:
          type:
          - string
          - 'null'
          format: uuid
        can_manage:
          type: boolean
          description: Whether this caller can change the explicit link policy; includes safe public/tag system
            roots for organization administrators.
...
    DelegatedInvite:
      type: object
      additionalProperties: false
      required:
      - operation_id
      - entry_id
      - invitation
      properties:
        operation_id:
          type: string
          format: uuid
        entry_id:
          type: string
          format: uuid
        invitation:
          $ref: '#/components/schemas/InvitationCreate'
    DelegatedLinkAccess:
      type: object
      additionalProperties: false
      required:
      - operation_id
      - entry_id
      - enabled
      properties:
        operation_id:
          type: string
          format: uuid
        entry_id:
          type: string
          format: uuid
        enabled:
          type: boolean
...